requirements.txt
onnx==1.7.0
onnxruntime==1.4.0
python_vlc==3.0.11115
httpx==0.14.1
uvicorn==0.11.8
face_recognition==1.3.0
gTTS==2.1.1
opencv_contrib_python==4.3.0.36
torch==1.6.0
pyttsx3==2.90
numpy==1.19.1
fastapi==0.61.0
PyYAML==5.3.1
